Features:   -functionality to trade resources (requires research)   -convert all reasonable values from double to int   -slightly randomize income   -rework blueprint cost boundaries to avoid resulting in wonky stats
Research work:
  -new research to unlock additional +1 research choices
  -remove meaningless _researchCur
  -give different research options different change weights based on:
    *how long since its been an option
	*how long since researched
	*if an upgrade or unlock
	*stored progress
  -rework research tree (again)
  -reduce increase in research cost and increase less for upgrades rather than unlocks

UI work:
  -optimize drawing all enemy attacks, only show numbers if zoomed in enough to reasonably read
  -use letters instead of numbers for blueprints to help differentiate from mech unit numbers
  -hide columns in build form / attack grid until needed

Many minor fixes and rebalances:
  -constructor no longer starts with repair
  -enemies dont throw away to expensive blueprints, they just wait
  -fix upgrade randomization bugs
  -fix builder bug where different implementation can have different ranges
  -core shield bug

Clean up and prioritize todo list

+++ WinFormsApp1/ClassLibrary1/IResearch.cs
@@ -9,10 +9,10 @@ namespace ClassLibrary1
     interface IResearch
     {
         Game Game { get; }
-        double GetLevel();
+        int GetLevel();
         Research.Type GetType();
-        double GetMinCost();
-        double GetMaxCost();
+        int GetMinCost();
+        int GetMaxCost();
         bool MakeType(Research.Type type);
         double GetMult(Research.Type type, double pow);
     }
